Drop nanopb int_size from hall effect fields
The config JSON layer in config_utils.cpp is macro-generated from each field's
protobuf type, so a UINT32 field is always handed to helpers taking uint32_t --
by value for serialization and by non-const reference for deserialization. A
narrowed C type cannot bind to either, which is why no field in this proto uses
(nanopb).int_size.

Using IS_8 on the rapid trigger percentages and IS_16 on the profile action array
therefore broke the build with ambiguous appendAsString() overloads and reference
binding failures across every translation unit that includes config.pb.h.

Costs about 390 bytes: Config_size goes 22791 -> 23178 of 32768.

+++ src/addons/he_trigger.cpp
@@ -530,7 +530,7 @@ void HETriggerAddon::applyCalibration(uint8_t actuationPercent, uint8_t pressPer
         int32_t deadzone = (span > 0) ? ((noise * 100) / span) : 3;
         if (deadzone < 2) deadzone = 2;
         if (deadzone > 10) deadzone = 10;
-        trigger.travelDeadzone = (uint8_t)deadzone;
+        trigger.travelDeadzone = (uint32_t)deadzone;
         trigger.has_travelDeadzone = true;
 
         trigger.actuationPoint = actuationPercent;
